Move utilIdTypes to config/entities
Caleydo/cohort#693
