+++ src/app/CoralApp.ts
@@ -28,7 +28,7 @@ import deleteModal from './templates/DeleteModal.html';
 import welcomeHtml from './templates/Welcome.html';
 import { getAnimatedLoadingText, handleDataLoadError, log } from '../util';
 import { CohortSelectionEvent, ConfirmTaskEvent, CONFIRM_TASK_EVENT_TYPE, PreviewConfirmEvent } from '../base/events';
-import { idCellline, idCovid19, idStudent, idTissue, IEntitySourceConfig } from '../utilIdTypes';
+import { idCellline, idCovid19, idStudent, idTissue, IEntitySourceConfig } from '../config/entities';
 import { niceName } from '../utilLabels';
 import { IDatasetDesc, IPanelDesc } from './interfaces';
 import { CohortSelectionListener } from './CoralSelectionListener';
